requirements.txt
# Reel Automation Pipeline Dependencies

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Async HTTP client
aiohttp>=3.9.0

# Data validation
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Database (PostgreSQL)
asyncpg>=0.28.0
sqlalchemy[asyncio]>=2.0.0

# Image processing
Pillow>=10.0.0

# CLI interface
click>=8.0.0

# Progress bars
tqdm>=4.65.0
